grading homework

+++ upload.php
@@ -29,26 +29,31 @@ if(isset($_POST["submit"]) && !empty($_FILES["file"]["name"])){
          // Allow certain file formats
     $allowTypes = array('jpg','png','jpeg','gif','pdf');
     if(in_array($fileType, $allowTypes)){
-        // Upload file to server
-        if(move_uploaded_file($_FILES["file"]["tmp_name"], $targetFilePath)){
+  
+        if(!file_exists($targetFilePath)){
+            // Upload file to server
+            if(move_uploaded_file($_FILES["file"]["tmp_name"], $targetFilePath)){
+                
             
-        
-            $query = $db->pdo->prepare('INSERT INTO student_assignment (student_id, title, description, semester, professor_assignment_id, file) VALUES (:student_id, :title, :description, :semester, :professor_assignment_id, :file)');
-            $query->bindParam(':student_id', $_SESSION['user_id']);
-            $query->bindParam(':title', $_POST['title']);
-            $query->bindParam(':description', $_POST['description']);
-            $query->bindParam(':semester', $_POST['selectSemester']);
-            $query->bindParam(':professor_assignment_id', $_POST['selectAssignment']);
-            $query->bindParam(':file', $fileName);
+                $query = $db->pdo->prepare('INSERT INTO student_assignment (student_id, title, description, semester, professor_assignment_id, file) VALUES (:student_id, :title, :description, :semester, :professor_assignment_id, :file)');
+                $query->bindParam(':student_id', $_SESSION['user_id']);
+                $query->bindParam(':title', $_POST['title']);
+                $query->bindParam(':description', $_POST['description']);
+                $query->bindParam(':semester', $_POST['selectSemester']);
+                $query->bindParam(':professor_assignment_id', $_POST['selectAssignment']);
+                $query->bindParam(':file', $fileName);
 
-            $query->execute();
-            if($query){
-                $statusMsg = "The file ".$fileName. " has been uploaded successfully.";
+                $query->execute();
+                if($query){
+                    $statusMsg = "The file ".$fileName. " has been uploaded successfully.";
+                }else{
+                    $statusMsg = "File upload failed, please try again.";
+                } 
             }else{
-                $statusMsg = "File upload failed, please try again.";
-            } 
+                $statusMsg = "Sorry, there was an error uploading your file.";
+            }
         }else{
-            $statusMsg = "Sorry, there was an error uploading your file.";
+            $statusMsg = "A file with that name already exists";
         }
     }else{
         $statusMsg = 'Sorry, only JPG, JPEG, PNG, GIF, & PDF files are allowed to upload.';
